Update flac to 1.5.0
CMake 4.0 removed compatibility with versions older than 3.5, which
would break building the vendored flac and ogg libraries were it not for
the workaround implemented in #40.

The root cause of this issue has been fixed in the upstream repositories
for FLAC and ogg, but only FLAC has released a new version (1.5.0) with
a fix.

* Update flac to 1.5.0
* Regenerate bindings using the bindgen script
* Suppress clippy lints on generated bindings